refactor(services): migrate comment_service to shared/comment
- Move comment_service.py → shared/comment.py
- Update import in task/show.py

Part of issue #2586: Phase 3-cleanup
